lambda_report/requirements.txt
boto3==1.28.85
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.24.3
